// extensions/category-tier-discount-native/generated/api.ts
export type Maybe<T> = T | null;
// ...
export type BuyerIdentity = {
  __typename?: 'BuyerIdentity';
  /** The [customer](https://help.shopify.com/manual/customers/manage-customers) that's interacting with the cart. */
  customer?: Maybe<Customer>;
  /** The email address of the customer that's interacting with the cart. */
  email?: Maybe<Scalars['String']['output']>;
  /**
   * Whether the customer is authenticated through their
   * [customer account](https://help.shopify.com/manual/customers/customer-accounts).
   */
  isAuthenticated: Scalars['Boolean']['output'];
  /** The phone number of the customer that's interacting with the cart. */
  phone?: Maybe<Scalars['String']['output']>;
  /**
   * The company of a B2B customer that's interacting with the cart.
   * Used to manage and track purchases made by businesses rather than individual customers.
   */
  purchasingCompany?: Maybe<PurchasingCompany>;
};

/**
 * The cart where the Function is running. A cart contains the merchandise that a customer intends to purchase
 * and information about the customer, such as the customer's email address and phone number.
 */
export type Cart = HasMetafields & {
  __typename?: 'Cart';
  /**
   * The custom attributes associated with a cart to store additional information. Cart attributes
   * allow you to collect specific information from customers on the **Cart** page, such as order notes,
   * gift wrapping requests, or custom product details. Attributes are stored as key-value pairs.
   */
  attribute?: Maybe<Attribute>;
  /**
   * Information about the customer that's interacting with the cart. It includes details such as the
   * customer's email and phone number, and the total amount of money the customer has spent in the store.
   * This information helps personalize the checkout experience and ensures that accurate pricing and delivery options
   * are displayed to customers.
   */
  buyerIdentity?: Maybe<BuyerIdentity>;
  /**
   * A breakdown of the costs that the customer will pay at checkout. It includes the total amount,
   * the subtotal before taxes and duties, the tax amount, and duty charges.
   */
  cost: CartCost;
  /** The items in a cart that are eligible for fulfillment and can be delivered to the customer. */
  deliverableLines: Array<DeliverableCartLine>;
  /**
   * A collection of items that are grouped by shared delivery characteristics. Delivery groups streamline
   * fulfillment by organizing items that can be shipped together, based on the customer's
   * shipping address. For example, if a customer orders a t-shirt and a pair of shoes that can be shipped
   * together, then the items are included in the same delivery group.
   *
   * In the [Order Discount](https://shopify.dev/docs/api/functions/reference/order-discounts) and
   * [Product Discount](https://shopify.dev/docs/api/functions/reference/product-discounts) legacy APIs,
   * the `cart.deliveryGroups` input is always an empty array. This means you can't access delivery groups when
   * creating Order Discount or Product Discount Functions. If you need to apply discounts to shipping costs,
   * then use the [Discount Function API](https://shopify.dev/docs/api/functions/reference/discount)
   * instead.
   */
  deliveryGroups: Array<CartDeliveryGroup>;
  /**
   * The items in a cart that the customer intends to purchase. A cart line is an entry in the
   * customer's cart that represents a single unit of a product variant. For example, if a customer adds two
   * different sizes of the same t-shirt to their cart, then each size is represented as a separate cart line.
   */
  lines: Array<CartLine>;
  /**
   * The additional fields on the **Cart** page that are required for international orders in specific countries,
   * such as customs information or tax identification numbers.
   */
  localizedFields: Array<LocalizedField>;
  /**
   * A [custom field](https://shopify.dev/docs/apps/build/custom-data) that stores additional information
   * about a Shopify resource, such as products, orders, and
   * [many more](https://shopify.dev/docs/api/admin-graphql/latest/enums/MetafieldOwnerType).
   * Using [metafields with Shopify Functions](https://shopify.dev/docs/apps/build/functions/input-output/metafields-for-input-queries)
   * enables you to customize the checkout experience.
   */
  metafield?: Maybe<Metafield>;
  /** The physical location where a retail order is created or completed. */
  retailLocation?: Maybe<Location>;
};
// ...
export type CartLinesDiscountsGenerateRunResult = {
  /** The list of operations to apply discounts to the cart. */
  operations: Array<CartOperation>;
};
// ...
export type CustomProduct = {
  __typename?: 'CustomProduct';
  /** Whether the merchandise is a gift card. */
  isGiftCard: Scalars['Boolean']['output'];
  /**
   * Whether the item needs to be shipped to the customer. For example, a
   * digital gift card doesn't need to be shipped, but a t-shirt does
   * need to be shipped.
   */
  requiresShipping: Scalars['Boolean']['output'];
  /**
   * The localized name for the product that displays to customers. The title is used to construct the product's
   * handle, which is a unique, human-readable string of the product's title. For example, if a product is titled
   * "Black Sunglasses", then the handle is `black-sunglasses`.
   */
  title: Scalars['String']['output'];
  /** The product variant's weight, in the system of measurement set in the `weightUnit` field. */
  weight?: Maybe<Scalars['Float']['output']>;
  /** The unit of measurement for weight. */
  weightUnit: WeightUnit;
};

/**
 * Represents a [customer](https://help.shopify.com/manual/customers/manage-customers).
 * `Customer` returns data including the customer's contact information and order history.
 */
export type Customer = HasMetafields & {
  __typename?: 'Customer';
  /**
   * The total amount that the customer has spent on orders.
   * The amount is converted from the shop's currency to the currency of the cart using a market rate.
   */
  amountSpent: MoneyV2;
  /**
   * The full name of the customer, based on the values for `firstName` and `lastName`.
   * If `firstName` and `lastName` aren't specified, then the value is the customer's email address.
   * If the email address isn't specified, then the value is the customer's phone number.
   */
  displayName: Scalars['String']['output'];
  /** The customer's email address. */
  email?: Maybe<Scalars['String']['output']>;
  /** The customer's first name. */
  firstName?: Maybe<Scalars['String']['output']>;
  /**
   * Whether the customer is associated with any of the specified tags. The customer must have at least one tag
   * from the list to return `true`.
   */
  hasAnyTag: Scalars['Boolean']['output'];
  /** Whether the customer is associated with the specified tags. */
  hasTags: Array<HasTagResponse>;
  /**
   * A [globally-unique ID](https://shopify.dev/docs/api/usage/gids)
   * for the customer.
   */
  id: Scalars['ID']['output'];
  /** The customer's last name. */
  lastName?: Maybe<Scalars['String']['output']>;
  /**
   * A [custom field](https://shopify.dev/docs/apps/build/custom-data) that stores additional information
   * about a Shopify resource, such as products, orders, and
   * [many more](https://shopify.dev/docs/api/admin-graphql/latest/enums/MetafieldOwnerType).
   * Using [metafields with Shopify Functions](https://shopify.dev/docs/apps/build/functions/input-output/metafields-for-input-queries)
   * enables you to customize the checkout experience.
   */
  metafield?: Maybe<Metafield>;
  /** The total number of orders that the customer has made at the store. */
  numberOfOrders: Scalars['Int']['output'];
};
// ...
export type Input = {
  __typename?: 'Input';
  /**
   * The cart where the Function is running. A cart contains the merchandise that a customer intends to purchase
   * and information about the customer, such as the customer's email address and phone number.
   */
  cart: Cart;
  /**
   * The discount node that owns the [Shopify
   * Function](https://shopify.dev/docs/apps/build/functions). Discounts are a way
   * for merchants to promote sales and special offers, or as customer loyalty
   * rewards. A single discount can be automatic or code-based, and can be applied
   * to a cart lines, orders, and delivery.
   */
  discount: Discount;
  /**
   * Discount codes entered by the buyer at checkout, excluding gift cards. For
   * cart.lines.discounts.generate.run and cart.delivery-options.discounts.generate.run targets, these
   * discount codes are validated to ensure they are not deleted, maintain an active status, and are eligible
   * for the current cart. For fetch targets, all entered discount codes are included, excluding gift cards.
   */
  enteredDiscountCodes: Array<EnteredDiscountCode>;
  /**
   * The result of the fetch target. Refer to [network access](https://shopify.dev/apps/build/functions/input-output/network-access/graphql)
   * for Shopify Functions. This input is only available in the
   * `cart.lines.discounts.generate.run` and
   * `cart.delivery-options.discounts.generate.run` extension targets.
   */
  fetchResult?: Maybe<HttpResponse>;
  /**
   * The regional and language settings that determine how the Function
   * handles currency, numbers, dates, and other locale-specific values
   * during discount calculations. These settings are based on the store's configured
   * [localization practices](https://shopify.dev/docs/apps/build/functions/localization-practices-shopify-functions).
   */
  localization: Localization;
  /**
   * The exchange rate used to convert discounts between the shop's default
   * currency and the currency that displays to the customer during checkout.
   * For example, if a store operates in USD but a customer is viewing discounts in EUR,
   * then the presentment currency rate handles this conversion for accurate pricing.
   */
  presentmentCurrencyRate: Scalars['Decimal']['output'];
  /**
   * Information about the shop where the Function is running, including the shop's timezone
   * setting and associated [metafields](https://shopify.dev/docs/apps/build/custom-data).
   */
  shop: Shop;
  /**
   * The discount code entered by a customer, which caused the [Discount Function](https://shopify.dev/docs/apps/build/discounts#build-with-shopify-functions) to run.
   * This input is only available in the `cart.lines.discounts.generate.run` and
   * `cart.delivery-options.discounts.generate.run` extension targets.
   */
  triggeringDiscountCode?: Maybe<Scalars['String']['output']>;
};
// ...
export type Metafield = {
  __typename?: 'Metafield';
  /** The data that's stored in the metafield, using JSON format. */
  jsonValue: Scalars['JSON']['output'];
  /**
   * The [type of data](https://shopify.dev/apps/metafields/types) that the metafield stores in
   * the `value` field.
   */
  type: Scalars['String']['output'];
  /**
   * The data that's stored in the metafield. The data is always stored as a string,
   * regardless of the [metafield's type](https://shopify.dev/apps/metafields/types).
   */
  value: Scalars['String']['output'];
};
// ...
export enum ProductDiscountSelectionStrategy {
  /** Apply all the discount candidates to eligible cart lines. This doesn't override discount combination or stacking rules. */
  All = 'ALL',
  /** Apply the first discount candidate to cart lines that satisfies conditions. */
  First = 'FIRST',
  /** Apply the discount to the cart line that offers the maximum reduction. */
  Maximum = 'MAXIMUM'
}
// ...
export type CartInput = { __typename?: 'Input', cart: { __typename?: 'Cart', buyerIdentity?: { __typename?: 'BuyerIdentity', customer?: { __typename?: 'Customer', metafield?: { __typename?: 'Metafield', value: string } | null } | null } | null, lines: Array<{ __typename?: 'CartLine', id: string, merchandise: { __typename: 'CustomProduct' } | { __typename: 'ProductVariant', product: { __typename?: 'Product', ipad: boolean, mac: boolean, accessories: boolean, iphone: boolean, dynamicCollections: Array<{ __typename?: 'CollectionMembership', collectionId: string, isMember: boolean }> } } }> }, discount: { __typename?: 'Discount', discountClasses: Array<DiscountClass>, discountConfig?: { __typename?: 'Metafield', value: string } | null, automaticConfig?: { __typename?: 'Metafield', value: string } | null } };